MIPS: BMIPS: Avoid referencing CKSEG1
bmips_smp_movevec() references the CKSEG1 constant, which is about to be
updated in order to support processors that might enable eXtended
KSEG0/1. In doing so, we will generate a reference to a function, which
is obviously not permissible within assembly. Fortunately,
bmips_smp_movevec() is only used on BMIPS4350 which does not support
eXtended KSEG0/1.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>