net: dsa: b53: Add support for interrupt controller
Add the ability to create an interrupt domain that is local to the switch and
which allows obtaining per-port link status changes. In case an external
interrupt line to the host is not wired, or the chip is too old, we poll the
link status register for link status changes which saves a bunch of MDIO
transactions.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>