net: dsa: bcm_sf2: Enable management mode
Now that we have all the necessary plumbing in place to get notified
when a multicast MAC address must be programmed, configure the switch to
operate in managed mode and let the network stack learn about management
traffic.

During suspend, we need to bring the switch back into unmanaged/dumb
forwarding mode to allow stations to communicate with each other despite
the host CPU being suspended.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>