net: dsa: Add ability to elect CPU port
In a configuration where multiple CPU ports are declared within the
platform configuration, it may be desirable to make sure that a
particular CPU port gets used. This is particularly true for Broadcom
switch that are fairly flexible to some extent in which port can be the
CPU port, yet will be more featureful if port 8 is elected.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>