soc: bcm: brcmstb: Add support for the "powercycle" reboot type
On chips where the BCM59400 PMIC is used, it is desirable for customers
to be able to issue a complete chip power cycle, which can be achieved
by requesting the PMIC to do that.

When the "powercycle" command is specified in the reboot arguments, we
translate that into a PSCI vendor specific reset and the reset type set
to 1.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>